+++ HISTORY.rst
@@ -6,6 +6,10 @@ Release History
 0.9.9
 +++++++++++++++
 
+IoT dps updates
+    * Introduces 'az iot dps compute-device-key' preview command group to generate derived device SAS key
+
+
 IoT Central updates
 
 * Current release involves re-grouping of IoT central commands.
@@ -15,7 +19,7 @@ IoT Central updates
   * Note all commands listed below are in preview
 
     * Introduces 'az iot central diagnostics' preview command group to perform application and device level diagnostics
-
+    * Introduces 'az iot central device compute-device-key' preview command group to generate derived device SAS key
       * 'az iot central app device registration-summary' moved to 'az iot central diagnostics registration-summary'
       * 'az iot central app monitor-properties' moved to 'az iot central diagnostics monitor-properties'
       * 'az iot central app validate-messages' moved to 'az iot central diagnostics validate-messages'